Extract git patch executor
