+++ docs/DECISIONS.md
@@ -224,3 +224,4 @@
 - GitHub pull request execution lives in `GitHubPullRequestToolExecutor.swift`, and raw git/gh process launching lives in `GitProcessRunner.swift`. `GitToolExecutor` remains the compatibility facade used by the tool router and remote planners, but PR command construction, PR-specific input validation, URL artifact extraction, and raw `Process` management should stay out of that facade.
 - Git worktree execution lives in `GitWorktreeToolExecutor.swift`. Worktree sibling-path validation, registered-worktree lookup, create/remove command construction, and worktree artifact reporting should stay beside that executor; `GitToolExecutor` remains a compatibility facade for existing tool-router call sites.
 - Shared git input validation lives in `GitInputValidator.swift`. Focused executors and SSH Remote git planning should use that neutral helper for trimming, git-name validation, and local relative-path validation instead of reaching back into `GitToolExecutor`; the facade may keep delegating compatibility wrappers only.
+- Git hunk patch execution lives in `GitPatchToolExecutor.swift`, and git tool errors live in `GitToolError.swift`. Hunk staging/restoring, patch temp-file handling, and patch-path metadata parsing should stay beside the patch executor; local and SSH Remote hunk planning should reuse `GitPatchToolExecutor.mismatchedPatchPath` instead of adding patch parsers to facades or remote planners.
